refactor: modularize loadout tray and campaign instance service contracts
