+++ src/services/persistence/CampaignInstanceService.ts
@@ -21,6 +21,12 @@ import type {
   IUnitDamageState,
 } from '../../types/campaign/CampaignInstanceInterfaces';
 import type { IPilotSkills } from '../../types/pilot/PilotInterfaces';
+import type {
+  ICampaignInstanceService,
+  IInstanceQueryOptions,
+  IUpdatePilotInstanceInput,
+  IUpdateUnitInstanceInput,
+} from './CampaignInstanceService.types';
 
 import {
   createUnitInstance,
@@ -34,117 +40,12 @@ import {
 } from '../../types/campaign/CampaignInterfaces';
 import { getIndexedDBService, STORES } from './IndexedDBService';
 
-// =============================================================================
-// Types
-// =============================================================================
-
-/**
- * Query options for listing instances
- */
-export interface IInstanceQueryOptions {
-  /** Filter by campaign ID */
-  readonly campaignId?: string;
-  /** Filter by status */
-  readonly status?: CampaignUnitStatus | CampaignPilotStatus;
-  /** Filter by force ID */
-  readonly forceId?: string;
-  /** Include only available (deployable) instances */
-  readonly availableOnly?: boolean;
-}
-
-/**
- * Unit instance update input
- */
-export interface IUpdateUnitInstanceInput {
-  readonly status?: CampaignUnitStatus;
-  readonly damageState?: IUnitDamageState;
-  readonly assignedPilotInstanceId?: string | null;
-  readonly forceId?: string | null;
-  readonly forceSlot?: number | null;
-  readonly totalKills?: number;
-  readonly missionsParticipated?: number;
-  readonly estimatedRepairCost?: number;
-  readonly estimatedRepairTime?: number;
-  readonly notes?: string;
-}
-
-/**
- * Pilot instance update input
- */
-export interface IUpdatePilotInstanceInput {
-  readonly status?: CampaignPilotStatus;
-  readonly currentSkills?: IPilotSkills;
-  readonly wounds?: number;
-  readonly currentXP?: number;
-  readonly campaignXPEarned?: number;
-  readonly killCount?: number;
-  readonly missionsParticipated?: number;
-  readonly assignedUnitInstanceId?: string | null;
-  readonly recoveryTime?: number;
-  readonly notes?: string;
-}
-
-// =============================================================================
-// Service Interface
-// =============================================================================
-
-/**
- * Campaign Instance Service interface
- */
-export interface ICampaignInstanceService {
-  // Unit Instances
-  createUnitInstance(
-    input: ICreateUnitInstanceInput,
-    unitData: {
-      version: number;
-      name: string;
-      chassis: string;
-      variant: string;
-      damageState?: IUnitDamageState;
-    },
-  ): Promise<ICampaignUnitInstance>;
-  getUnitInstance(id: string): Promise<ICampaignUnitInstance | undefined>;
-  updateUnitInstance(
-    id: string,
-    updates: IUpdateUnitInstanceInput,
-  ): Promise<ICampaignUnitInstance>;
-  deleteUnitInstance(id: string): Promise<void>;
-  listUnitInstances(
-    options?: IInstanceQueryOptions,
-  ): Promise<ICampaignUnitInstance[]>;
-
-  // Pilot Instances
-  createPilotInstanceFromVault(
-    input: ICreatePilotInstanceFromVaultInput,
-    pilotData: {
-      name: string;
-      callsign?: string;
-      skills: IPilotSkills;
-    },
-  ): Promise<ICampaignPilotInstance>;
-  createPilotInstanceFromStatblock(
-    input: ICreatePilotInstanceFromStatblockInput,
-  ): Promise<ICampaignPilotInstance>;
-  getPilotInstance(id: string): Promise<ICampaignPilotInstance | undefined>;
-  updatePilotInstance(
-    id: string,
-    updates: IUpdatePilotInstanceInput,
-  ): Promise<ICampaignPilotInstance>;
-  deletePilotInstance(id: string): Promise<void>;
-  listPilotInstances(
-    options?: IInstanceQueryOptions,
-  ): Promise<ICampaignPilotInstance[]>;
-
-  // Assignments
-  assignPilotToUnit(
-    pilotInstanceId: string,
-    unitInstanceId: string,
-  ): Promise<void>;
-  unassignPilot(pilotInstanceId: string): Promise<void>;
-
-  // Bulk Operations
-  deleteInstancesForCampaign(campaignId: string): Promise<void>;
-}
+export type {
+  ICampaignInstanceService,
+  IInstanceQueryOptions,
+  IUpdatePilotInstanceInput,
+  IUpdateUnitInstanceInput,
+} from './CampaignInstanceService.types';
 
 // =============================================================================
 // IndexedDB Implementation
